add -merge- option to collapse
with 1mm obs, `collapse .. , merge` takes 50% of the time of egen even
for the simplest case of a sum over one variable.

For more variables or more complex ops, this scales even faster

Note: any advantage of this (hash+join) approach is completely lost
(over the sort+merge approach) if the data is already sorted by the by()
vars!